docs/requirements.txt
Sphinx==5.3.0
sphinx-rtd-theme
sphinx_autoapi
PyGithub
GitPython
loguru

